restructure the code and split service.py file
