+++ src/ensembl/production/metadata/grpc/service.py
@@ -1,0 +1,32 @@
+#  See the NOTICE file distributed with this work for additional information
+#  regarding copyright ownership.
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#      http://www.apache.org/licenses/LICENSE-2.0
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+from concurrent import futures
+import grpc
+import logging
+
+from ensembl.production.metadata.grpc import ensembl_metadata_pb2_grpc
+from ensembl.production.metadata.grpc.servicer import EnsemblMetadataServicer
+
+
+def serve():
+    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
+    ensembl_metadata_pb2_grpc.add_EnsemblMetadataServicer_to_server(
+        EnsemblMetadataServicer(), server
+    )
+    server.add_insecure_port("[::]:50051")
+    server.start()
+    server.wait_for_termination()
+
+
+if __name__ == "__main__":
+    logging.basicConfig()
+    serve()
