+++ src/ensembl/production/metadata/grpc/utils.py
@@ -10,18 +10,17 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 import itertools
-from concurrent import futures
-import grpc
-import logging
-import sqlalchemy as db
+from ensembl.production.metadata.grpc import ensembl_metadata_pb2
 
-from ensembl.production.metadata import ensembl_metadata_pb2_grpc
-from ensembl.production.metadata import ensembl_metadata_pb2
-
-from ensembl.production.metadata.config import MetadataConfig as cfg
+from ensembl.production.metadata.grpc.config import MetadataConfig as cfg
 
 from ensembl.production.metadata.api.genome import GenomeAdaptor
 from ensembl.production.metadata.api.release import ReleaseAdaptor
+
+from ensembl.production.metadata.grpc.protobuf_msg_factory import create_genome, create_karyotype, \
+    create_top_level_statistics, create_top_level_statistics_by_uuid, create_assembly, create_species, \
+    create_sub_species, create_genome_uuid, create_datasets, create_genome_sequence, create_release, \
+    create_dataset_infos, populate_dataset_info
 
 
 def connect_to_db():
@@ -234,16 +233,6 @@ def get_genome_by_name(db_conn, ensembl_name, site_name, release_version):
     return create_genome()
 
 
-def populate_dataset_info(data):
-    return ensembl_metadata_pb2.DatasetInfos.DatasetInfo(
-        dataset_uuid=data.Dataset.dataset_uuid,
-        dataset_name=data.Dataset.name,
-        dataset_version=data.Dataset.version,
-        dataset_label=data.Dataset.label,
-        version=int(data.EnsemblRelease.version),
-    )
-
-
 def get_datasets_list_by_uuid(db_conn, genome_uuid, release_version=0):
     if genome_uuid is None:
         return create_datasets()
@@ -335,294 +324,3 @@ def get_dataset_by_genome_id(db_conn, genome_uuid, requested_dataset_type):
         dataset_type=requested_dataset_type
     )
     return create_dataset_infos(genome_uuid, requested_dataset_type, dataset_results)
-
-
-def create_species(species_data=None, taxo_info=None):
-    if species_data is None:
-        return ensembl_metadata_pb2.Species()
-
-    species = ensembl_metadata_pb2.Species(
-        genome_uuid=species_data.Genome.genome_uuid,
-        taxon_id=species_data.Organism.taxonomy_id,
-        scientific_name=species_data.Organism.scientific_name,
-        scientific_parlance_name=species_data.Organism.scientific_parlance_name,
-        ncbi_common_name=taxo_info["ncbi_common_name"],
-        common_name=taxo_info["common_name"],
-        alternative_names=taxo_info["alternative_names"],
-    )
-    return species
-
-
-def create_top_level_statistics(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.TopLevelStatistics()
-    species = ensembl_metadata_pb2.TopLevelStatistics(
-        organism_uuid=data["organism_uuid"],
-        statistics=data["statistics"],
-    )
-    return species
-
-
-def create_top_level_statistics_by_uuid(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.TopLevelStatisticsByUUID()
-
-    species = ensembl_metadata_pb2.TopLevelStatisticsByUUID(
-        genome_uuid=data["genome_uuid"],
-        statistics=data["statistics"],
-    )
-    return species
-
-
-def create_karyotype(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.Karyotype()
-
-    karyotype = ensembl_metadata_pb2.Karyotype(
-        genome_uuid=data.Genome.genome_uuid,
-        code=data.Assembly.level,
-        chromosomal=str(data.AssemblySequence.chromosomal),
-        location=data.AssemblySequence.sequence_location,
-    )
-    return karyotype
-
-
-def create_sub_species(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.SubSpecies()
-    sub_species = ensembl_metadata_pb2.SubSpecies(
-        organism_uuid=data["organism_uuid"],
-        species_name=data["species_name"],
-        species_type=data["species_type"],
-    )
-    return sub_species
-
-
-def create_assembly(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.AssemblyInfo()
-
-    assembly = ensembl_metadata_pb2.AssemblyInfo(
-        assembly_uuid=data.Assembly.assembly_uuid,
-        accession=data.Assembly.accession,
-        level=data.Assembly.level,
-        name=data.Assembly.name,
-        chromosomal=data.AssemblySequence.chromosomal,
-        length=data.AssemblySequence.length,
-        sequence_location=data.AssemblySequence.sequence_location,
-        sequence_checksum=data.AssemblySequence.sequence_checksum,
-        ga4gh_identifier=data.AssemblySequence.ga4gh_identifier,
-    )
-    return assembly
-
-
-def create_genome_uuid(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.GenomeUUID()
-
-    genome_uuid = ensembl_metadata_pb2.GenomeUUID(
-        genome_uuid=data["genome_uuid"]
-    )
-    return genome_uuid
-
-
-def create_genome(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.Genome()
-
-    assembly = ensembl_metadata_pb2.Assembly(
-        accession=data.Assembly.accession,
-        name=data.Assembly.name,
-        ucsc_name=data.Assembly.ucsc_name,
-        level=data.Assembly.level,
-        ensembl_name=data.Assembly.ensembl_name,
-    )
-
-    taxon = ensembl_metadata_pb2.Taxon(
-        taxonomy_id=data.Organism.taxonomy_id,
-        scientific_name=data.Organism.scientific_name,
-        strain=data.Organism.strain,
-    )
-    # TODO: fetch common_name(s) from ncbi_taxonomy database
-
-    organism = ensembl_metadata_pb2.Organism(
-        display_name=data.Organism.display_name,
-        strain=data.Organism.strain,
-        scientific_name=data.Organism.scientific_name,
-        url_name=data.Organism.url_name,
-        ensembl_name=data.Organism.ensembl_name,
-        scientific_parlance_name=data.Organism.scientific_parlance_name,
-        organism_uuid=data.Organism.organism_uuid,
-    )
-
-    release = ensembl_metadata_pb2.Release(
-        release_version=data.EnsemblRelease.version,
-        release_date=str(data.EnsemblRelease.release_date),
-        release_label=data.EnsemblRelease.label,
-        is_current=data.EnsemblRelease.is_current,
-        site_name=data.EnsemblSite.name,
-        site_label=data.EnsemblSite.label,
-        site_uri=data.EnsemblSite.uri,
-    )
-
-    genome = ensembl_metadata_pb2.Genome(
-        genome_uuid=data.Genome.genome_uuid,
-        created=str(data.Genome.created),
-        assembly=assembly,
-        taxon=taxon,
-        organism=organism,
-        release=release,
-    )
-
-    return genome
-
-
-def create_genome_sequence(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.GenomeSequence()
-
-    genome_sequence = ensembl_metadata_pb2.GenomeSequence(
-        accession=data.AssemblySequence.accession,
-        name=data.AssemblySequence.name,
-        sequence_location=data.AssemblySequence.sequence_location,
-        length=data.AssemblySequence.length,
-        chromosomal=data.AssemblySequence.chromosomal
-    )
-
-    return genome_sequence
-
-
-def create_release(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.Release()
-
-    release = ensembl_metadata_pb2.Release(
-        release_version=data.EnsemblRelease.version,
-        release_date=str(data.EnsemblRelease.release_date),
-        release_label=data.EnsemblRelease.label,
-        is_current=data.EnsemblRelease.is_current,
-        site_name=data.EnsemblSite.name,
-        site_label=data.EnsemblSite.label,
-        site_uri=data.EnsemblSite.uri
-    )
-
-    return release
-
-
-def create_datasets(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.Datasets()
-
-    return ensembl_metadata_pb2.Datasets(
-        genome_uuid=data["genome_uuid"], datasets=data["datasets"]
-    )
-
-
-def create_dataset_info(data=None):
-    if data is None:
-        return ensembl_metadata_pb2.DatasetInfos.DatasetInfo()
-
-    return ensembl_metadata_pb2.DatasetInfos.DatasetInfo(
-        dataset_uuid=data.Dataset.dataset_uuid,
-        dataset_name=data.Dataset.name,
-        name=data.Attribute.name,
-        type=data.Attribute.type,
-        dataset_version=data.Dataset.version,
-        dataset_label=data.Dataset.label,
-        version=int(data.EnsemblRelease.version),
-        value=data.DatasetAttribute.value,
-    )
-
-
-def create_dataset_infos(genome_uuid=None, requested_dataset_type=None, data=None):
-    if data is None or data == []:
-        return ensembl_metadata_pb2.DatasetInfos()
-
-    dataset_infos = [create_dataset_info(result) for result in data]
-    return ensembl_metadata_pb2.DatasetInfos(
-        genome_uuid=genome_uuid,
-        dataset_type=requested_dataset_type,
-        dataset_infos=dataset_infos,
-    )
-
-
-class EnsemblMetadataServicer(ensembl_metadata_pb2_grpc.EnsemblMetadataServicer):
-    def __init__(self):
-        self.db = connect_to_db()
-
-    def GetSpeciesInformation(self, request, context):
-        return get_species_information(self.db, request.genome_uuid)
-
-    def GetAssemblyInformation(self, request, context):
-        return get_assembly_information(self.db, request.assembly_uuid)
-
-    def GetGenomesByAssemblyAccessionID(self, request, context):
-        return get_genomes_from_assembly_accession_iterator(
-            self.db, request.assembly_accession
-        )
-
-    def GetSubSpeciesInformation(self, request, context):
-        return get_sub_species_info(self.db, request.organism_uuid)
-
-    def GetKaryotypeInformation(self, request, context):
-        return get_karyotype_information(self.db, request.genome_uuid)
-
-    def GetTopLevelStatistics(self, request, context):
-        return get_top_level_statistics(self.db, request.organism_uuid)
-
-    def GetTopLevelStatisticsByUUID(self, request, context):
-        return get_top_level_statistics_by_uuid(self.db, request.genome_uuid)
-
-    def GetGenomeUUID(self, request, context):
-        return get_genome_uuid(self.db, request.ensembl_name, request.assembly_name)
-
-    def GetGenomeByUUID(self, request, context):
-        return get_genome_by_uuid(self.db, request.genome_uuid, request.release_version)
-
-    def GetGenomesByKeyword(self, request, context):
-        return get_genomes_by_keyword_iterator(
-            self.db, request.keyword, request.release_version
-        )
-
-    def GetGenomeByName(self, request, context):
-        return get_genome_by_name(
-            self.db, request.ensembl_name, request.site_name, request.release_version
-        )
-
-    def GetRelease(self, request, context):
-        return release_iterator(
-            self.db, request.site_name, request.release_version, request.current_only
-        )
-
-    def GetReleaseByUUID(self, request, context):
-        return release_by_uuid_iterator(self.db, request.genome_uuid)
-
-    def GetGenomeSequence(self, request, context):
-        return genome_sequence_iterator(
-            self.db, request.genome_uuid, request.chromosomal_only
-        )
-
-    def GetDatasetsListByUUID(self, request, context):
-        return get_datasets_list_by_uuid(
-            self.db, request.genome_uuid, request.release_version
-        )
-
-    def GetDatasetInformation(self, request, context):
-        return get_dataset_by_genome_id(
-            self.db, request.genome_uuid, request.dataset_type
-        )
-
-
-def serve():
-    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
-    ensembl_metadata_pb2_grpc.add_EnsemblMetadataServicer_to_server(
-        EnsemblMetadataServicer(), server
-    )
-    server.add_insecure_port("[::]:50051")
-    server.start()
-    server.wait_for_termination()
-
-
-if __name__ == "__main__":
-    logging.basicConfig()
-    serve()
